Extract and reuse test setup
